@@ -4,10 +4,10 @@ name = "ccproxy"
 dynamic = ["version"]
 description = "API server that provides an Anthropic and OpenAI compatible interface over Claude Code, allowing to use your Claude OAuth account or over the API."
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dependencies = [
-  "check-jsonschema>=0.33.2",
-  "claude-code-sdk>=0.0.14", # Using GitHub version with PR #60 fix
+  "aiosqlite>=0.21.0",
+  "jsonschema>=0.33.2",
   "fastapi[standard]>=0.115.14",
   "httpx>=0.28.1",
   "httpx-sse>=0.4.1",
