@@ -1,5 +1,4 @@
 from ._version import __version__
-from .main import app
 
 
-__all__ = ["__version__", "app"]
+__all__ = ["__version__"]
