@@ -1,5 +1,14 @@
-from .app import app, app_main, main, version_callback
-from .commands.claude import claude
+from .commands.serve import api, claude
+from .helpers import get_rich_toolkit
+from .main import app, app_main, main, version_callback
 
 
-__all__ = ["app", "main", "version_callback", "claude", "app_main"]
+__all__ = [
+    "app",
+    "main",
+    "version_callback",
+    "api",
+    "claude",
+    "app_main",
+    "get_rich_toolkit",
+]
