@@ -1,9 +1,8 @@
 """Command modules for Claude Code Proxy API CLI."""
 
-from .api import api
 from .auth import app as auth_app
-from .claude import claude
 from .config import app as config_app
+from .serve import api
 
 
-__all__ = ["api", "auth_app", "claude", "config_app"]
+__all__ = ["api", "auth_app", "config_app"]
