@@ -1,13 +1,13 @@
 """Docker path management with clean API."""
 
-import logging
 from pathlib import Path
 from typing import Self
 
 from pydantic import BaseModel, field_validator
+from structlog import get_logger
 
 
-logger = logging.getLogger(__name__)
+logger = get_logger(__name__)
 
 
 class DockerPath(BaseModel):
@@ -116,7 +116,7 @@ class DockerPathSet:
         """
         self.base_host_path = Path(base_host_path).resolve() if base_host_path else None
         self.paths: dict[str, DockerPath] = {}
-        self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+        self.logger = get_logger(f"{__name__}.{self.__class__.__name__}")
 
     def add(
         self, name: str, container_path: str, host_subpath: str | None = None
