@@ -2,6 +2,8 @@
 
 from typing import Literal, TypeAlias
 
+from typing_extensions import TypedDict
+
 
 # Message and content types
 MessageRole: TypeAlias = Literal["user", "assistant", "system", "tool"]
@@ -16,7 +18,7 @@ OpenAIContentType: TypeAlias = Literal["text", "image_url"]
 # Tool-related types
 ToolChoiceType: TypeAlias = Literal["auto", "any", "tool", "none", "required"]
 OpenAIToolChoiceType: TypeAlias = Literal["none", "auto", "required"]
-ToolType: TypeAlias = Literal["function"]
+ToolType: TypeAlias = Literal["function", "custom"]
 
 # Response format types
 ResponseFormatType: TypeAlias = Literal["text", "json_object", "json_schema"]
@@ -77,3 +79,24 @@ OpenAIObjectType: TypeAlias = Literal[
 
 # Permission behavior types
 PermissionBehavior: TypeAlias = Literal["allow", "deny"]
+
+
+# Usage and streaming related types
+class UsageData(TypedDict, total=False):
+    """Token usage data extracted from streaming or non-streaming responses."""
+
+    input_tokens: int | None
+    output_tokens: int | None
+    cache_read_input_tokens: int | None
+    cache_creation_input_tokens: int | None
+    event_type: StreamEventType | None
+
+
+class StreamingTokenMetrics(TypedDict, total=False):
+    """Accumulated token metrics during streaming."""
+
+    tokens_input: int | None
+    tokens_output: int | None
+    cache_read_tokens: int | None
+    cache_write_tokens: int | None
+    cost_usd: float | None
